requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-extensions==4.1
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework-stubs==3.16.4
greenlet==3.2.4
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.44
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
